Combines Netbackup backup/restore test suite into main backup suite
This is part of the effort to get all backup/restore tests using the
same test suite. Since the Netbackup tests take significantly longer, we
only run a subset of the regular test suite. We also tag scenarios to
allow parallel runs on separate hosts in CI.

This suite will take 1h, 40 mins after parallelization, down from the
current 2h, 20 mins.
