+++ gpMgmt/test/behave/mgmt_utils/restores.feature
@@ -11,7 +11,7 @@ Feature: Validate command line arguments
         Given the backup test is initialized with no backup files
         And database "testdb" exists
         And there is a "heap" table "public.table1" in "testdb" with data
-        When the user runs command "gpcrondump -a -x testdb"
+        When the user runs "gpcrondump -a -x testdb"
         And the timestamp from gpcrondump is stored
         When the user runs gpdbrestore -e with the stored timestamp and options "--table-file test/behave/mgmt_utils/steps/data/special_chars/funny_char_table.txt"
         Then gpdbrestore should return a return code of 2
@@ -22,7 +22,7 @@ Feature: Validate command line arguments
         When the user runs gpdbrestore -e with the stored timestamp and options "--redirect A\\t\\n.,!1"
         Then gpdbrestore should return a return code of 2
         And gpdbrestore should print "Name has an invalid character" to stdout
-        When the user runs command "gpdbrestore -s "A\\t\\n.,!1""
+        When the user runs "gpdbrestore -s "A\\t\\n.,!1""
         Then gpdbrestore should return a return code of 2
         And gpdbrestore should print "Name has an invalid character" to stdout
         When the user runs gpdbrestore -e with the stored timestamp and options "-T public.table1 --change-schema A\\t\\n.,!1"
@@ -65,3 +65,20 @@ Feature: Validate command line arguments
         When the user runs gpdbrestore without -e with the stored timestamp and options "-T public.ao_index_table"
         Then gpdbrestore should return a return code of 0
         And gpdbrestore should print "Restore type               = Incremental Table Restore" to stdout
+
+    Scenario: Valid option combinations for gpdbrestore
+        When the user runs "gpdbrestore -t 20140101010101 --truncate -a"
+        Then gpdbrestore should return a return code of 2
+        And gpdbrestore should print "--truncate can be specified only with -S, -T, or --table-file option" to stdout
+        When the user runs "gpdbrestore -t 20140101010101 --truncate -e -T public.foo -a"
+        Then gpdbrestore should return a return code of 2
+        And gpdbrestore should print "Cannot specify --truncate and -e together" to stdout
+        And there is a table-file "/tmp/table_file_foo" with tables "public.ao_table, public.co_table"
+        And the user runs "gpdbrestore -t 20140101010101 -T public.ao_table --table-file /tmp/table_file_foo"
+        Then gpdbrestore should return a return code of 2
+        And gpdbrestore should print "Cannot specify -T and --table-file together" to stdout
+        Then the file "/tmp/table_file_foo" is removed from the system
+        When the user runs "gpdbrestore -u /tmp --ddboost -s bkdb"
+        Then gpdbrestore should return a return code of 2
+        And gpdbrestore should print "-u cannot be used with DDBoost parameters" to stdout
+
